requirements.txt
pyyaml
duckdb>=1.3  # older (<1.3) versions of httpsfs are not so good at retrying
pyarrow
surt
pytest
pytest-sugar

# development
flake8
